// src/enum/flat/flatTables.enum.ts
export enum FlatTableList {
  costcenter = 'costcenter',
  account_assignment_category = 'account_assignment_category',
  address = 'address',
  business_area = "business_area",
  company = 'company',
  currency = 'currency',
  department = 'DEPARTMENT',
  gl_code = 'gl_code',
  industry_category = 'industry_category',
  internal_order_number = 'internal_order_number',
  payment_terms = 'payment_terms',
  plant = 'plant',
  profit_center = 'profit_center',
  purchase_org = 'purchase_org',
  sap_class = 'sap_class',
  sap_model = 'sap_model',
  sap_series = 'sap_series',
  supplier_tag = 'supplier_tag',
  uom = 'uom',
  user_roles = 'user_roles',
  warranty = 'warranty',
  previousPurchase = 'previousPurchase',
  supplier = "SUPPLIER",
  partNumber = "Part_Number",
  decimal = "decimal",
  event_type = 'event_type'
}